Reorganize ensemble into subpackage and fix imports
