+++ mdpow/analysis/__init__.py
@@ -1,0 +1,2 @@
+# Analysis Subpackage
+# Alia Lescoulie (ALescoulie) 2021
